+++ Assets/Scripts/Scriptables/Health/ObjectHealth.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "ObjectHealth", menuName = "Scriptables/Health/ObjectHealth")]
 public class ObjectHealth : Health
 {
-    public override void TakeDamage(GameObject obj, int currentHealth)
+    public override void TakeDamage(GameObject obj, float currentHealth)
     {
         if (currentHealth <= 0)
         {
