+++ Assets/Scripts/Scriptables/Health/PlayerHealth.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "PlayerHealth", menuName = "Scriptables/Health/PlayerHealth")]
 public class PlayerHealth : Health
 {
-    public override void TakeDamage(GameObject obj, int currentHealth)
+    public override void TakeDamage(GameObject obj, float currentHealth)
     {
         if (currentHealth <= 0)
         {
